pyproject.toml
[project]
name = "anthropic-engineering-ja"
version = "0.1.0"
description = "Anthropic Engineering ブログを日本語 PDF の読み物にするツール"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "markdownify>=0.12",
    "jinja2>=3.1",
    "markdown>=3.6",
    "pygments>=2.18",
    "python-frontmatter>=1.1",
    "playwright>=1.47",
]

[tool.uv]
package = false
